# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API OmniGest",
        "description": "Documentação da API Omnigest",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/centroCusto": {
            "get": {
                "tags": [
                    "Centro de Custo"
                ],
                "summary": "Lista todos os centros de custo",
                "operationId": "117db102afb1e6db772328e2cb537c49",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Quantidade de itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filtro para buscar centros de custo",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page_number",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de centros de custo retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CentroCusto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Centro de Custo"
                ],
                "summary": "Cria um novo centro de custo",
                "operationId": "73e3392daf0808573cba4979e94fb09f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "des_centro_custo_cco"
                                ],
                                "properties": {
                                    "des_centro_custo_cco": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Centro de Custo criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CentroCusto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/centroCusto/{id_centro_custo}": {
            "put": {
                "tags": [
                    "Centro de Custo"
                ],
                "summary": "Atualiza um centro de custo existente",
                "operationId": "753bc7830f2167bd1411fe6e93dcbf3d",
                "parameters": [
                    {
                        "name": "id_centro_custo",
                        "in": "path",
                        "description": "ID do centro de custo a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "des_centro_custo_cco"
                                ],
                                "properties": {
                                    "des_centro_custo_cco": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Centro de Custo atualizado com sucesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Centro de Custo"
                ],
                "summary": "Remove um centro de custo",
                "operationId": "a30890c05806c8fa92cd7215465d3c73",
                "parameters": [
                    {
                        "name": "id_centro_custo",
                        "in": "path",
                        "description": "ID do centro de custo a ser removido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Centro de Custo removido com sucesso"
                    }
                }
            }
        },
        "/cliente": {
            "post": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Cria um novo cliente",
                "operationId": "bc9a364274fd2f1b76621ace38ad1f6a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cliente"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/cliente/{id_cliente}": {
            "get": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Obtém um cliente pelo ID",
                "operationId": "35a75147a8e78e07829ee6b67574d3a8",
                "parameters": [
                    {
                        "name": "id_cliente",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Cliente não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Atualiza um cliente",
                "operationId": "b7cd6a484a355f0623534c9ac04869ef",
                "parameters": [
                    {
                        "name": "id_cliente",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cliente"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente atualizado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Deleta um cliente",
                "operationId": "fb97d010c130b3810dd52de8ecdf01be",
                "parameters": [
                    {
                        "name": "id_cliente",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente deletado com sucesso"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CentroCusto": {
                "properties": {
                    "id_centro_custo_cco": {
                        "type": "integer"
                    },
                    "id_empresa_cco": {
                        "description": "Chave estrangeira referenciando a empresa",
                        "type": "integer"
                    },
                    "des_centro_custo_cco": {
                        "type": "string"
                    },
                    "is_ativo_cco": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Cliente": {
                "properties": {
                    "id_cliente_cli": {
                        "type": "integer"
                    },
                    "id_empresa_cli": {
                        "description": "Chave estrangeira referenciando a empresa",
                        "type": "integer"
                    },
                    "des_cliente_cli": {
                        "type": "string"
                    },
                    "telefone_cliente_cli": {
                        "type": "string"
                    },
                    "email_cliente_cli": {
                        "type": "string",
                        "format": "email"
                    },
                    "documento_cliente_cli": {
                        "type": "string"
                    },
                    "endereco_cliente_cli": {
                        "type": "string"
                    },
                    "is_ativo_cli": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "id_centro_custo_cli": {
                        "description": "Chave estrangeira referenciando o centro de custo",
                        "type": "integer"
                    },
                    "id_origem_cliente_cli": {
                        "description": "Chave estrangeira referenciando a origem do cliente",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}